[Refactor] 스타일 폴더 정리 및 그에 따른 경로 수정 #590
