+++ components/modal/event/AnnouncementModal.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import { useRecoilState } from 'recoil';
+import { useSetRecoilState } from 'recoil';
 import { modalState } from 'utils/recoil/modal';
 import { Announcement } from 'types/modalTypes';
-import styles from 'styles/modal/AnnouncementModal.module.scss';
+import styles from 'styles/modal/event/AnnouncementModal.module.scss';
 
 type AnnouncementModalProps = {
   announcements: Announcement[];
@@ -10,7 +10,7 @@ type AnnouncementModalProps = {
 export default function AnnouncementModal({
   announcements,
 }: AnnouncementModalProps) {
-  const [modal, setModal] = useRecoilState(modalState);
+  const setModal = useSetRecoilState(modalState);
   const [neverSeeAgain, setNeverSeeAgain] = useState<boolean>(false);
   const onCheck = () => {
     setNeverSeeAgain((prev) => !prev);
@@ -27,24 +27,21 @@ export default function AnnouncementModal({
 
   return (
     <div className={styles.container}>
-      <div className={styles.announcementTitle}>Notice!</div>
-      <div className={styles.phrase}>
-        <div className={styles.emoji}></div>
-        <ul className={styles.announcementList}>
-          {announcements.map((el: Announcement, index) => {
-            return (
-              <li key={index}>
-                <div className={styles.title}>{el.title}</div>
-                <ul className={styles.content}>
-                  {el.content.map((e: string, idx) => (
-                    <li key={idx}>{e}</li>
-                  ))}
-                </ul>
-              </li>
-            );
-          })}
-        </ul>
-      </div>
+      <div className={styles.title}>Notice!</div>
+      <ul className={styles.announcementList}>
+        {announcements.map((el: Announcement, index) => {
+          return (
+            <li key={index}>
+              <div>{`<${el.title}>`}</div>
+              <ul className={styles.content}>
+                {el.content.map((e: string, idx) => (
+                  <li key={idx}>{e}</li>
+                ))}
+              </ul>
+            </li>
+          );
+        })}
+      </ul>
       <div>
         <input
           type='checkbox'
