+++ components/modal/match/MatchEnrollModal.tsx
@@ -5,7 +5,7 @@ import { reloadMatchState } from 'utils/recoil/match';
 import { errorState } from 'utils/recoil/error';
 import { modalState } from 'utils/recoil/modal';
 import instance from 'utils/axios';
-import styles from 'styles/modal/MatchEnrollModal.module.scss';
+import styles from 'styles/modal/match/MatchEnrollModal.module.scss';
 
 export default function MatchEnrollModal({
   slotId,
