+++ components/modal/menu/ReportModal.tsx
@@ -4,7 +4,7 @@ import { errorState } from 'utils/recoil/error';
 import { openMenuBarState } from 'utils/recoil/layout';
 import { modalState } from 'utils/recoil/modal';
 import instance from 'utils/axios';
-import styles from 'styles/modal/ReportModal.module.scss';
+import styles from 'styles/modal/menu/ReportModal.module.scss';
 
 interface Report {
   category: string;
