+++ components/modal/statChange/Celebration.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import styles from 'styles/modal/StatChangeModal.module.scss';
+import styles from 'styles/modal/afterGame/StatChangeModal.module.scss';
 
 export default function Celebration() {
   useEffect(() => {
