+++ styles/modal/afterGame/AfterGameModal.module.scss
