+++ styles/modal/afterGame/StatChangeModal.module.scss
