+++ styles/modal/event/AnnouncementModal.module.scss
@@ -3,42 +3,40 @@
 .container {
   @include modal-container;
   width: 80vw;
-  max-width: 20rem;
+  max-width: 15rem;
 }
 
-.announcementTitle {
-  color: white;
-  font-size: 2.5rem;
-  font-family: $title-font;
-  text-shadow: $text-shadow;
-  letter-spacing: 0.1rem;
+.title {
+  @include modalTitle;
+  font-size: 2rem;
 }
 
 .announcementList {
   list-style: none;
-  width: auto;
-  height: 60vh;
+  width: 100%;
   max-height: 60vh;
-  margin: 0 0.6rem 0.5rem;
   color: black;
-  font-size: 1.2rem;
-  font-weight: 700;
   overflow-y: scroll;
+  overflow-x: hidden;
 
   > li {
-    margin-top: 1.2rem;
-    margin-bottom: 1rem;
+    list-style: none;
+    margin-block: 1rem;
+    > div {
+      margin-left: 0.3rem;
+      font-size: 1.2rem;
+      font-weight: 700;
+    }
   }
 }
 
 .content {
-  margin-left: 0.8rem;
+  margin-top: 0.8rem;
   font-size: 0.9rem;
-  font-weight: 400;
 
   > li {
-    list-style: square;
-    margin: 0.4rem 0;
+    list-style: none;
+    margin-bottom: 0.3rem;
     white-space: pre-wrap;
   }
 }
