+++ styles/modal/event/WelcomeModal.module.scss
