+++ styles/modal/match/MatchManualModal.module.scss
@@ -7,11 +7,7 @@
 }
 
 .title {
-  color: white;
-  font-size: 2.5rem;
-  font-family: $title-font;
-  text-shadow: $text-shadow;
-  letter-spacing: 0.1rem;
+  @include modalTitle;
 }
 
 .toggleContainer {
