+++ styles/modal/match/MatchRejectModal.module.scss
